Reimplement default sample permissions settings UI
