+++ client/src/js/settings/actions.js
@@ -9,7 +9,7 @@
 
 import {
     GET_SETTINGS,
-    UPDATE_SETTING,
+    UPDATE_SETTINGS,
     GET_CONTROL_READAHEAD
 } from "../actionTypes";
 
@@ -22,11 +22,12 @@ export function getSettings () {
 export function updateSetting (key, value) {
     let update = {};
     update[key] = value;
+    return updateSettings(update);
+}
 
+export function updateSettings (update) {
     return {
-        type: UPDATE_SETTING.REQUESTED,
-        key,
-        value,
+        type: UPDATE_SETTINGS.REQUESTED,
         update
     };
 }
