+++ client/src/js/settings/reducers.js
@@ -7,10 +7,9 @@
  *
  */
 
-import { assign } from "lodash";
 import {
     GET_SETTINGS,
-    UPDATE_SETTING,
+    UPDATE_SETTINGS,
     GET_CONTROL_READAHEAD
 } from "../actionTypes";
 
@@ -26,26 +25,19 @@ const reducer = (state = initialState, action) => {
     switch (action.type) {
 
         case GET_SETTINGS.SUCCEEDED:
-            return assign({}, state, {data: action.data});
+            return {...state, data: action.data};
 
-        case UPDATE_SETTING.SUCCEEDED:
-            return assign({}, state, {data: assign({}, state.data, action.update)});
+        case UPDATE_SETTINGS.SUCCEEDED:
+            return {...state, data: {...state.data, ...action.update}};
 
         case GET_CONTROL_READAHEAD.REQUESTED:
-            return assign({}, state, {
-                readaheadPending: true
-            });
+            return {...state, readaheadPending: true};
 
         case GET_CONTROL_READAHEAD.SUCCEEDED:
-            return assign({}, state, {
-                readahead: action.data,
-                readaheadPending: false
-            });
+            return {...state, readahead: action.data, readaheadPending: false};
 
         case GET_CONTROL_READAHEAD.FAILED:
-            return assign({}, state, {
-                readaheadPending:false
-            });
+            return {...state, readaheadPending:false};
 
         default:
             return state;
