+++ client/src/js/settings/sagas.js
@@ -7,40 +7,40 @@
  *
  */
 
-import { call, put, throttle, takeEvery, takeLatest } from "redux-saga/effects";
+import { put, throttle, takeEvery, takeLatest } from "redux-saga/effects";
 
 import settingsAPI from "./api";
 import virusesAPI from "../viruses/api";
 import { setPending } from "../wrappers";
-import { GET_SETTINGS, UPDATE_SETTING, GET_CONTROL_READAHEAD } from "../actionTypes";
+import { GET_SETTINGS, UPDATE_SETTINGS, GET_CONTROL_READAHEAD } from "../actionTypes";
 
 export function* watchSettings () {
     yield takeLatest(GET_SETTINGS.REQUESTED, getSettings);
-    yield takeEvery(UPDATE_SETTING.REQUESTED, updateSetting);
+    yield takeEvery(UPDATE_SETTINGS.REQUESTED, updateSettings);
     yield throttle(120, GET_CONTROL_READAHEAD.REQUESTED, getControlReadahead);
 }
 
 function* getSettings () {
     try {
-        const response = yield call(settingsAPI.get);
+        const response = yield settingsAPI.get();
         yield put({type: GET_SETTINGS.SUCCEEDED, data: response.body});
     } catch (error) {
         yield put({type: GET_SETTINGS.FAILED}, error);
     }
 }
 
-function* updateSetting (action) {
+function* updateSettings (action) {
     yield setPending(function* () {
         try {
-            const response = yield call(settingsAPI.update, action.update);
+            const response = yield settingsAPI.update(action.update);
             yield put({
-                type: UPDATE_SETTING.SUCCEEDED,
+                type: UPDATE_SETTINGS.SUCCEEDED,
                 settings: response.body,
                 key: action.key,
                 update: action.update
             });
         } catch(error) {
-            yield put({type: UPDATE_SETTING.FAILED, key: action.key});
+            yield put({type: UPDATE_SETTINGS.FAILED, key: action.key});
         }
     }, action);
 }
